correct stderrors prepare blockchain test filler
