+++ retesteth/ethObjects/blockchainTest/scheme_blockchainTest.cpp
@@ -1,6 +1,22 @@
 #include "scheme_blockchainTest.h"
 using namespace test;
 
+
+scheme_blockchainTestBase::fieldChecker::fieldChecker(DataObject const& _test)
+{
+    requireJsonFields(_test, "blockchainTest " + _test.getKey(),
+        {{"_info", {{DataType::Object}, jsonField::Optional}},
+            {"blocks", {{DataType::Array}, jsonField::Required}},
+            {"expect", {{DataType::Array}, jsonField::Optional}},
+            {"genesisBlockHeader", {{DataType::Object}, jsonField::Required}},
+            {"genesisRLP", {{DataType::String}, jsonField::Optional}},
+            {"lastblockhash", {{DataType::String}, jsonField::Optional}},
+            {"network", {{DataType::String}, jsonField::Optional}},
+            {"postState", {{DataType::Object}, jsonField::Optional}},
+            {"pre", {{DataType::Object}, jsonField::Required}},
+            {"sealEngine", {{DataType::String}, jsonField::Optional}}});
+}
+
 scheme_blockchainTest::fieldChecker::fieldChecker(DataObject const& _test)
 {
     requireJsonFields(_test, "blockchainTest " + _test.getKey(),
@@ -15,12 +31,20 @@ scheme_blockchainTest::fieldChecker::fieldChecker(DataObject const& _test)
             {"sealEngine", {{DataType::String}, jsonField::Optional}}});
 }
 
-scheme_blockchainTest::scheme_blockchainTest(DataObject const& _test)
+scheme_blockchainTestBase::scheme_blockchainTestBase(DataObject const& _test)
   : object(_test),
     m_checker(_test),
     m_pre(_test.atKey("pre")),
-    m_post(_test.atKey("postState")),
     m_genesisHeader(_test.atKey("genesisBlockHeader"))
+{
+    if (_test.count("sealEngine"))
+        m_sealEngine = _test.atKey("sealEngine").asString();
+    else
+        m_sealEngine = "NoProof";
+}
+
+scheme_blockchainTest::scheme_blockchainTest(DataObject const& _test)
+  : scheme_blockchainTestBase(_test), m_checker(_test), m_post(_test.atKey("postState"))
 {
     for (auto const& data : _test.atKey("blocks").getSubObjects())
     {
@@ -29,17 +53,11 @@ scheme_blockchainTest::scheme_blockchainTest(DataObject const& _test)
             data.atKey("rlp").type() == DataType::String, "Block rlp field must be string!");
         m_blockRLPs.push_back(data.atKey("rlp").asString());
     }
-
-    if (_test.count("sealEngine"))
-        m_sealEngine = _test.atKey("sealEngine").asString();
-    else
-        m_sealEngine = "NoProof";
 }
 
-DataObject scheme_blockchainTest::getGenesisForRPC()
+DataObject scheme_blockchainTestBase::getGenesisForRPC()
 {
     DataObject genesis = prepareGenesisParams(getData().atKey("network").asString(), m_sealEngine);
-    // genesis["genesis"] = getEnv().getDataForRPC();
 
     DataObject data;
     data["author"] = m_genesisHeader.getData().atKey("coinbase");
