+++ retesteth/testSuites/StateTests.cpp
@@ -64,10 +64,10 @@ DataObject FillTestAsBlockchain(DataObject const& _testFile, TestSuite::TestSuit
 
     RPCSession& session = RPCSession::instance(TestOutputHelper::getThreadID());
     // run transactions on all networks that we need
-    for (auto const& net : test.getAllNetworksFromExpectSection())
+    for (auto const& net : test.getExpectSection().getAllNetworksFromExpectSection())
     {
         // run transactions for defined expect sections only
-        for (auto const& expect : test.getExpectSections())
+        for (auto const& expect : test.getExpectSection().getExpectSections())
         {
             // if expect section for this networks
             if (expect.getNetworks().count(net))
@@ -169,14 +169,14 @@ DataObject FillTest(DataObject const& _testFile, TestSuite::TestSuiteOptions& _o
     filledTest["transaction"] = test.getGenTransaction().getData();
 
     // run transactions on all networks that we need
-    for (auto const& net: test.getAllNetworksFromExpectSection())
+    for (auto const& net : test.getExpectSection().getAllNetworksFromExpectSection())
     {
         DataObject forkResults;
         forkResults.setKey(net);
         session.test_setChainParams(test.getGenesisForRPC(net, "NoReward").asJson());
 
         // run transactions for defined expect sections only
-        for (auto const& expect : test.getExpectSections())
+        for (auto const& expect : test.getExpectSection().getExpectSections())
         {
             // if expect section for this networks
             if (expect.getNetworks().count(net))
@@ -362,7 +362,7 @@ DataObject StateTestSuite::doTests(DataObject const& _input, TestSuiteOptions& _
 TestSuite::TestPath StateTestSuite::suiteFolder() const
 {
     if (Options::get().fillchain)
-        return TestSuite::TestPath(fs::path("BlockchainTests") / "GeneralStateTests2");
+        return TestSuite::TestPath(fs::path("BlockchainTests") / "GeneralStateTests");
     return TestSuite::TestPath(fs::path("GeneralStateTests"));
 }
